app/requirements.txt
flask
google-cloud-storage
google-auth